+++ build.gradle
@@ -31,6 +31,10 @@ dependencies {
     compile 'com.google.android.gms:play-services:5.0.89'
     compile 'com.astuetz:pagerslidingtabstrip:1.0.1'
     compile 'org.w3c:smil:1.0.0'
+    compile 'com.github.machinarius:preferencefragment:0.1.1'
+    compile ("com.doomonafireball.betterpickers:library:1.5.2") {
+            exclude group: 'com.android.support', module: 'support-v4'
+    }
 
     androidTestCompile 'com.squareup:fest-android:1.0.8'
